2025.06.23 size()の戻り値をsize_t型(uint)にしておく

+++ sample_20220425/Q.04/transformSparseMatrix.cpp
@@ -26,10 +26,10 @@ int main() {
     {0, 0, 0, 2, 0},
     {0, 0, 0, 0, 1}
   };
-  
+
   vector<vector<int>> output = transformSparseMatrix(input);
-  int rows = output.size();
-  int cols = output[0].size();
+  size_t rows = output.size();
+  size_t cols = output[0].size();
 
   cout << "[sparseMatrix]\n{";
   for (int i = 0; i < rows; i++) {
